Fix route resolution
