+++ api/app/controllers/Applications.scala
@@ -2,7 +2,7 @@ package controllers
 
 import io.apibuilder.api.v0.models.{ApplicationForm, MoveForm}
 import io.apibuilder.api.v0.models.json._
-import db.{Authorization, OrganizationsDao, ApplicationsDao}
+import db.{Authorization, OrganizationsDao, ApplicationsDao, VersionsDao}
 import javax.inject.{Inject, Singleton}
 import lib.Validation
 import play.api.mvc._
@@ -12,7 +12,8 @@ import java.util.UUID
 @Singleton
 class Applications @Inject() (
   applicationsDao: ApplicationsDao,
-  organizationsDao: OrganizationsDao
+  organizationsDao: OrganizationsDao,
+  versionsDao: VersionsDao
 ) extends Controller {
 
   def get(
@@ -131,4 +132,46 @@ class Applications @Inject() (
       }
     }
   }
+
+  def getMetadataAndVersionsByApplicationKey(
+    orgKey: String,
+    applicationKey: String,
+    limit: Long = 25,
+    offset: Long = 0
+  ) = AnonymousRequest { request =>
+    applicationsDao.findByOrganizationKeyAndApplicationKey(request.authorization, orgKey, applicationKey) match {
+      case None => NotFound
+      case Some(application) => {
+        val versions = versionsDao.findAllVersions(
+          request.authorization,
+          applicationGuid = Some(application.guid),
+          limit = limit,
+          offset = offset
+        )
+        Ok(Json.toJson(versions))
+      }
+    }
+  }
+
+  def getMetadataAndVersionsAndLatestTxtByApplicationKey(
+    orgKey: String,
+    applicationKey: String,
+    limit: Long = 25,
+    offset: Long = 0
+  ) = AnonymousRequest { request =>
+    applicationsDao.findByOrganizationKeyAndApplicationKey(request.authorization, orgKey, applicationKey) match {
+      case None => NotFound
+      case Some(application) => {
+        versionsDao.findAllVersions(
+          request.authorization,
+          applicationGuid = Some(application.guid),
+          limit = 1
+        ).headOption match {
+          case None => NotFound
+          case Some(v) => Ok(v.version)
+        }
+      }
+    }
+  }
+
 }
