+++ api/conf/routes
@@ -59,14 +59,14 @@ GET        /watches/:guid                                            controllers
 GET        /watches/check                                            controllers.Watches.getCheck(user_guid: _root_.scala.Option[_root_.java.util.UUID], organization_key: String, application_key: String)
 POST       /watches                                                  controllers.Watches.post()
 DELETE     /watches/:guid                                            controllers.Watches.deleteByGuid(guid: _root_.java.util.UUID)
+GET        /:orgKey/metadata/:applicationKey/versions                controllers.Applications.getMetadataAndVersionsByApplicationKey(orgKey: String, applicationKey: String, limit: Long ?= 25, offset: Long ?= 0)
+GET        /:orgKey/metadata/:applicationKey/versions/latest.txt     controllers.Applications.getMetadataAndVersionsAndLatestTxtByApplicationKey(orgKey: String, applicationKey: String)
 GET        /:orgKey                                                  controllers.Applications.get(orgKey: String, name: _root_.scala.Option[String], guid: _root_.scala.Option[_root_.java.util.UUID], key: _root_.scala.Option[String], has_version: _root_.scala.Option[Boolean], limit: Long ?= 25, offset: Long ?= 0)
 POST       /:orgKey                                                  controllers.Applications.post(orgKey: String)
 PUT        /:orgKey/:applicationKey                                  controllers.Applications.putByApplicationKey(orgKey: String, applicationKey: String)
 DELETE     /:orgKey/:applicationKey                                  controllers.Applications.deleteByApplicationKey(orgKey: String, applicationKey: String)
 POST       /:orgKey/:applicationKey/move                             controllers.Applications.postMoveByApplicationKey(orgKey: String, applicationKey: String)
 GET        /:orgKey/:applicationKey/:version/:generatorKey           controllers.Code.get(orgKey: String, applicationKey: String, version: String, generatorKey: String)
-GET        /:orgKey/metadata/:applicationKey/versions                controllers.Versions.getMetadataAndVersionsByApplicationKey(orgKey: String, applicationKey: String, limit: Long ?= 25, offset: Long ?= 0)
-GET        /:orgKey/metadata/:applicationKey/versions/latest.txt     controllers.Versions.getMetadataAndVersionsAndLatestTxtByApplicationKey(orgKey: String, applicationKey: String)
 GET        /:orgKey/:applicationKey                                  controllers.Versions.getByApplicationKey(orgKey: String, applicationKey: String, limit: Long ?= 25, offset: Long ?= 0)
 GET        /:orgKey/:applicationKey/:version                         controllers.Versions.getByApplicationKeyAndVersion(orgKey: String, applicationKey: String, version: String)
 POST       /:orgKey/:version                                         controllers.Versions.postByVersion(orgKey: String, version: String)
